Yet more test improvements
